break entity-reference cycle
