+++ src/platform/loader-base.ts
@@ -29,6 +29,7 @@ import '../runtime/type-from-literal.js';
 import '../runtime/handle-constructors.js';
 import '../runtime/noop-proxy.js';
 import '../runtime/storageNG/store-constructors.js';
+import '../runtime/entity-utils.js';
 
 type ParticleCtor = typeof Particle;
